Just ran perf. A few things pop out:
As expected, reads and writes to memory - even to the stack! - dominate the runtime of many functions according to perf annotate and perf stat (67% of the time is spent stalling). Also, div is super expensive apparently, and explains why ualign_up/down are taking up 9% of the total runtime.

Of the nalloc functions, the biggest culprit is arena_init, with 13% of total
sample occurence. perf reveals that the most commonly sampled instruction
(60%) inside of it is the very first write, which is an int write write to
arena->size. It could simply be an issue of too many inits, since I'm being
very greedy with arenas now, but also note that this is the sharing test and
that other cores will be writing to wayward_blocks, which is on the same cache
line as arena->size. Could this be false sharing?

Result: no, running with just 1 thread shows the same result, with arena_init taking up 15% of the time (more than the mmap responsible for actually allocating the arena to init, which is at 14%).

New hypothesis: mmap is being lazy/overcommital in allocating frames. The write in arena->init is page faulting and triggerring the actual allocation inside the kernel. I can test this with an mlock after mmap, or perhaps with a syscall tracer (which would be interesting anyway).